Import and standardize scRNAseq data across different formats

Automatically infers data format of scRNAseq object, or a path to that object. It then uses the appropriate functions to import that data and convert it to a SingleCellExperiment, which is recognized by other EWCE functions.

ingest_data(
  obj,
  input_type = "guess",
  output_type = "SingleCellExperiment",
  custom_reader = NULL,
  save_dir = tempdir(),
  filename = "scKirby_output",
  save_output = T,
  overwrite = F,
  return_filepath = F,
  verbose = T,
  ...
)

Arguments

obj
Single-cell data object or path to saved single-cell data.

input_type
Format of obj. By default, the type will be inferred.

output_type
Format to convert obj to.

custom_reader
Custom function to read obj into R.

save_dir
Directory to save the converted obj.

filename
Name to save the converted obj.

save_output
Whether or not to save the converted obj.

overwrite
If a file of the same name exists, overwrite it.

return_filepath
If TRUE, a list with both the converted object and the saved file path will be returned (instead of just the converted object).

...
Additional arguments to be passed to scKirby::read_data().

Source

SeuratDisk anndata (R) anndata (python) loomR SingleCellExperiment DelayedArray workshop zellkonverter

Examples